Resolved merge conflicts.
