+++ platform-sdk/platform-apps/tests/MigrationTestingTool/src/main/java/com/swirlds/demo/migration/MigrationTestingToolState.java
@@ -252,110 +252,4 @@ public class MigrationTestingToolState extends PlatformMerkleStateRoot {
     public int getMinimumSupportedVersion() {
         return ClassVersion.VIRTUAL_MAP;
     }
-
-
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void init(
-            @NonNull final Platform platform,
-            @NonNull final InitTrigger trigger,
-            @Nullable final SoftwareVersion previousSoftwareVersion) {
-        super.init(platform, trigger, previousSoftwareVersion);
-
-        final MerkleMap<AccountID, MapValue> merkleMap = getMerkleMap();
-        if (merkleMap != null) {
-            logger.info(STARTUP.getMarker(), "MerkleMap initialized with {} values", merkleMap.size());
-        }
-        final VirtualMap<?, ?> virtualMap = getVirtualMap();
-        if (virtualMap != null) {
-            logger.info(STARTUP.getMarker(), "VirtualMap initialized with {} values", virtualMap.size());
-        }
-        selfId = platform.getSelfId();
-
-        if (trigger == InitTrigger.GENESIS) {
-            logger.warn(STARTUP.getMarker(), "InitTrigger was {} when expecting RESTART or RECONNECT", trigger);
-        }
-
-        if (previousSoftwareVersion == null || previousSoftwareVersion.compareTo(PREVIOUS_SOFTWARE_VERSION) != 0) {
-            logger.warn(
-                    STARTUP.getMarker(),
-                    "previousSoftwareVersion was {} when expecting it to be {}",
-                    previousSoftwareVersion,
-                    PREVIOUS_SOFTWARE_VERSION);
-        }
-
-        if (trigger == InitTrigger.GENESIS) {
-            logger.info(STARTUP.getMarker(), "Doing genesis initialization");
-            genesisInit(platform);
-        }
-    }
-
-    @Override
-    public void preHandle(
-            final @NonNull Event event,
-            final @NonNull Consumer<ScopedSystemTransaction<StateSignatureTransaction>> stateSignatureTransaction) {
-        event.forEachTransaction(transaction -> {
-
-            // We don't want to consume deprecated EventTransaction.STATE_SIGNATURE_TRANSACTION system transactions in
-            // the callback, since it's intended to be used only
-            // for the new form of encoded system transactions in Bytes.
-            // We skip the current iteration, if it processes a deprecated system transaction with the
-            // EventTransaction.STATE_SIGNATURE_TRANSACTION type.
-            if (transaction.isSystem()) {
-                return;
-            }
-
-            if (isSystemTransaction(transaction.getApplicationTransaction())) {
-                consumeSystemTransaction(transaction, event, stateSignatureTransaction);
-            }
-        });
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void handleConsensusRound(
-            @NonNull final Round round,
-            @NonNull final PlatformStateModifier platformState,
-            @NonNull final Consumer<ScopedSystemTransaction<StateSignatureTransaction>> stateSignatureTransaction) {
-        throwIfImmutable();
-        for (final Iterator<ConsensusEvent> eventIt = round.iterator(); eventIt.hasNext(); ) {
-            final ConsensusEvent event = eventIt.next();
-            for (final Iterator<ConsensusTransaction> transIt = event.consensusTransactionIterator();
-                 transIt.hasNext(); ) {
-                final ConsensusTransaction trans = transIt.next();
-                if (trans.isSystem()) {
-                    continue;
-                }
-                if (isSystemTransaction(trans.getApplicationTransaction())) {
-                    consumeSystemTransaction(trans, event, stateSignatureTransaction);
-                    continue;
-                }
-
-                final MigrationTestingToolTransaction mTrans =
-                        TransactionUtils.parseTransaction(trans.getApplicationTransaction());
-                mTrans.applyTo(this);
-            }
-        }
-    }
-
-
-    private void consumeSystemTransaction(
-            final @NonNull Transaction transaction,
-            final @NonNull Event event,
-            final @NonNull Consumer<ScopedSystemTransaction<StateSignatureTransaction>>
-                            stateSignatureTransactionCallback) {
-        try {
-            final var stateSignatureTransaction =
-                    StateSignatureTransaction.PROTOBUF.parse(transaction.getApplicationTransaction());
-            stateSignatureTransactionCallback.accept(new ScopedSystemTransaction<>(
-                    event.getCreatorId(), event.getSoftwareVersion(), stateSignatureTransaction));
-        } catch (final ParseException e) {
-            logger.error("Failed to parse StateSignatureTransaction", e);
-        }
-    }
 }
