+++ platform-sdk/platform-apps/tests/StressTestingTool/src/main/java/com/swirlds/demo/stress/StressTestingToolState.java
@@ -73,138 +73,6 @@ public class StressTestingToolState extends PlatformMerkleStateRoot {
      * {@inheritDoc}
      */
     @Override
-    public void init(
-            @NonNull final Platform platform,
-            @NonNull final InitTrigger trigger,
-            @Nullable final SoftwareVersion previousSoftwareVersion) {
-        super.init(platform, trigger, previousSoftwareVersion);
-
-        this.config = platform.getContext().getConfiguration().getConfigData(StressTestingToolConfig.class);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void preHandle(
-            @NonNull final Event event,
-            @NonNull
-                    final Consumer<ScopedSystemTransaction<StateSignatureTransaction>>
-                            stateSignatureTransactionCallback) {
-        event.forEachTransaction(transaction -> {
-            // We are not interested in pre-handling any system transactions, as they are
-            // specific for the platform only.We also don't want to consume deprecated
-            // EventTransaction.STATE_SIGNATURE_TRANSACTION system transactions in the
-            // callback,since it's intended to be used only for the new form of encoded system
-            // transactions in Bytes.Thus, we can directly skip the current
-            // iteration, if it processes a deprecated system transaction with the
-            // EventTransaction.STATE_SIGNATURE_TRANSACTION type.
-            if (transaction.isSystem()) {
-                return;
-            }
-
-            if (areTransactionBytesSystemOnes(transaction)) {
-                consumeSystemTransaction(transaction, event, stateSignatureTransactionCallback);
-            }
-        });
-
-        busyWait(config.preHandleTime());
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void handleConsensusRound(
-            @NonNull final Round round,
-            @NonNull final PlatformStateModifier platformState,
-            @NonNull
-                    final Consumer<ScopedSystemTransaction<StateSignatureTransaction>>
-                            stateSignatureTransactionCallback) {
-        throwIfImmutable();
-
-        for (final var event : round) {
-            event.consensusTransactionIterator().forEachRemaining(transaction -> {
-                // We are not interested in handling any system transactions, as they are
-                // specific for the platform only.We also don't want to consume deprecated
-                // EventTransaction.STATE_SIGNATURE_TRANSACTION system transactions in the
-                // callback,since it's intended to be used only for the new form of encoded system
-                // transactions in Bytes.Thus, we can directly skip the current
-                // iteration, if it processes a deprecated system transaction with the
-                // EventTransaction.STATE_SIGNATURE_TRANSACTION type.
-                if (transaction.isSystem()) {
-                    return;
-                }
-
-                if (areTransactionBytesSystemOnes(transaction)) {
-                    consumeSystemTransaction(transaction, event, stateSignatureTransactionCallback);
-                } else {
-                    handleTransaction(transaction);
-                }
-            });
-        }
-    }
-
-    private void handleTransaction(@NonNull final ConsensusTransaction trans) {
-        runningSum +=
-                ByteUtils.byteArrayToLong(trans.getApplicationTransaction().toByteArray(), 0);
-        busyWait(config.handleTime());
-    }
-
-    /**
-     * Checks if the transaction bytes are system ones. The test creates application transactions with max length of 4.
-     * System transactions will be always bigger than that.
-     *
-     * @param transaction the consensus transaction to check
-     * @return true if the transaction bytes are system ones, false otherwise
-     */
-    private boolean areTransactionBytesSystemOnes(@NonNull final Transaction transaction) {
-        final var transactionBytes = transaction.getApplicationTransaction();
-
-        if (transactionBytes.length() == 0) {
-            return false;
-        }
-
-        return transactionBytes.getByte(0) != APPLICATION_TRANSACTION_MARKER;
-    }
-
-    /**
-     * Converts a transaction to a {@link StateSignatureTransaction} and then consumes it into a callback.
-     *
-     * @param transaction the transaction to consume
-     * @param event the event that contains the transaction
-     * @param stateSignatureTransactionCallback the callback to call with the system transaction
-     */
-    private void consumeSystemTransaction(
-            final @NonNull Transaction transaction,
-            final @NonNull Event event,
-            final @NonNull Consumer<ScopedSystemTransaction<StateSignatureTransaction>>
-                            stateSignatureTransactionCallback) {
-        try {
-            final var stateSignatureTransaction =
-                    StateSignatureTransaction.PROTOBUF.parse(transaction.getApplicationTransaction());
-            stateSignatureTransactionCallback.accept(new ScopedSystemTransaction<>(
-                    event.getCreatorId(), event.getSoftwareVersion(), stateSignatureTransaction));
-        } catch (final com.hedera.pbj.runtime.ParseException e) {
-            logger.error("Failed to parse StateSignatureTransaction", e);
-        }
-    }
-
-    @SuppressWarnings("all")
-    private void busyWait(@NonNull final Duration duration) {
-        if (!duration.isZero() && !duration.isNegative()) {
-            final long start = System.nanoTime();
-            final long nanos = duration.toNanos();
-            while (System.nanoTime() - start < nanos) {
-                // busy wait
-            }
-        }
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
     public long getClassId() {
         return CLASS_ID;
     }
