+++ platform-sdk/platform-apps/tests/StressTestingTool/src/main/java/com/swirlds/demo/stress/StressTestingToolStateLifecycles.java
@@ -16,6 +16,8 @@
 
 package com.swirlds.demo.stress;
 
+import static com.swirlds.demo.stress.TransactionPool.APPLICATION_TRANSACTION_MARKER;
+
 import com.hedera.hapi.platform.event.StateSignatureTransaction;
 import com.swirlds.common.context.PlatformContext;
 import com.swirlds.common.utility.ByteUtils;
@@ -28,12 +30,16 @@ import com.swirlds.platform.system.SoftwareVersion;
 import com.swirlds.platform.system.address.AddressBook;
 import com.swirlds.platform.system.events.Event;
 import com.swirlds.platform.system.transaction.ConsensusTransaction;
+import com.swirlds.platform.system.transaction.Transaction;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.Nullable;
 import java.time.Duration;
 import java.util.function.Consumer;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 public class StressTestingToolStateLifecycles implements StateLifecycles<StressTestingToolState> {
+    private static final Logger logger = LogManager.getLogger(StressTestingToolStateLifecycles.class);
 
     /** supplies the app config */
     private StressTestingToolConfig config;
@@ -52,6 +58,23 @@ public class StressTestingToolStateLifecycles implements StateLifecycles<StressT
             @NonNull Event event,
             @NonNull StressTestingToolState state,
             @NonNull Consumer<ScopedSystemTransaction<StateSignatureTransaction>> stateSignatureTransactionCallback) {
+        event.forEachTransaction(transaction -> {
+            // We are not interested in pre-handling any system transactions, as they are
+            // specific for the platform only.We also don't want to consume deprecated
+            // EventTransaction.STATE_SIGNATURE_TRANSACTION system transactions in the
+            // callback,since it's intended to be used only for the new form of encoded system
+            // transactions in Bytes.Thus, we can directly skip the current
+            // iteration, if it processes a deprecated system transaction with the
+            // EventTransaction.STATE_SIGNATURE_TRANSACTION type.
+            if (transaction.isSystem()) {
+                return;
+            }
+
+            if (areTransactionBytesSystemOnes(transaction)) {
+                consumeSystemTransaction(transaction, event, stateSignatureTransactionCallback);
+            }
+        });
+
         busyWait(config.preHandleTime());
     }
 
@@ -72,7 +95,26 @@ public class StressTestingToolStateLifecycles implements StateLifecycles<StressT
             @NonNull StressTestingToolState state,
             @NonNull Consumer<ScopedSystemTransaction<StateSignatureTransaction>> stateSignatureTransactionCallback) {
         state.throwIfImmutable();
-        round.forEachTransaction(v -> handleTransaction(v, state));
+        for (final var event : round) {
+            event.consensusTransactionIterator().forEachRemaining(transaction -> {
+                // We are not interested in handling any system transactions, as they are
+                // specific for the platform only.We also don't want to consume deprecated
+                // EventTransaction.STATE_SIGNATURE_TRANSACTION system transactions in the
+                // callback,since it's intended to be used only for the new form of encoded system
+                // transactions in Bytes.Thus, we can directly skip the current
+                // iteration, if it processes a deprecated system transaction with the
+                // EventTransaction.STATE_SIGNATURE_TRANSACTION type.
+                if (transaction.isSystem()) {
+                    return;
+                }
+
+                if (areTransactionBytesSystemOnes(transaction)) {
+                    consumeSystemTransaction(transaction, event, stateSignatureTransactionCallback);
+                } else {
+                    handleTransaction(transaction, state);
+                }
+            });
+        }
     }
 
     private void handleTransaction(@NonNull final ConsensusTransaction trans, StressTestingToolState state) {
@@ -101,5 +143,44 @@ public class StressTestingToolStateLifecycles implements StateLifecycles<StressT
     @Override
     public void onNewRecoveredState(@NonNull StressTestingToolState recoveredState) {
         // no-op
+    }
+
+    /**
+     * Checks if the transaction bytes are system ones. The test creates application transactions with max length of 4.
+     * System transactions will be always bigger than that.
+     *
+     * @param transaction the consensus transaction to check
+     * @return true if the transaction bytes are system ones, false otherwise
+     */
+    private boolean areTransactionBytesSystemOnes(@NonNull final Transaction transaction) {
+        final var transactionBytes = transaction.getApplicationTransaction();
+
+        if (transactionBytes.length() == 0) {
+            return false;
+        }
+
+        return transactionBytes.getByte(0) != APPLICATION_TRANSACTION_MARKER;
+    }
+
+    /**
+     * Converts a transaction to a {@link StateSignatureTransaction} and then consumes it into a callback.
+     *
+     * @param transaction the transaction to consume
+     * @param event the event that contains the transaction
+     * @param stateSignatureTransactionCallback the callback to call with the system transaction
+     */
+    private void consumeSystemTransaction(
+            final @NonNull Transaction transaction,
+            final @NonNull Event event,
+            final @NonNull Consumer<ScopedSystemTransaction<StateSignatureTransaction>>
+                            stateSignatureTransactionCallback) {
+        try {
+            final var stateSignatureTransaction =
+                    StateSignatureTransaction.PROTOBUF.parse(transaction.getApplicationTransaction());
+            stateSignatureTransactionCallback.accept(new ScopedSystemTransaction<>(
+                    event.getCreatorId(), event.getSoftwareVersion(), stateSignatureTransaction));
+        } catch (final com.hedera.pbj.runtime.ParseException e) {
+            logger.error("Failed to parse StateSignatureTransaction", e);
+        }
     }
 }
